fix: cgpoint x, y的类型纠正

+++ lib/src/type/platform/ios_type/cg_point.dart
@@ -2,11 +2,11 @@ import 'package:foundation_fluttify/foundation_fluttify.dart';
 import 'package:foundation_fluttify/src/constants.dart';
 
 class CGPoint extends Ref {
-  Future<int> get x {
+  Future<double> get x {
     return kMethodChannel.invokeMethod('CGPoint::getX', {'refId': refId});
   }
 
-  Future<int> get y {
+  Future<double> get y {
     return kMethodChannel.invokeMethod('CGPoint::getY', {'refId': refId});
   }
 }
